+++ packages/easy-sitemap/__tests__/index.test.ts
@@ -39,5 +39,58 @@ describe('Basic Sitemap Tests', () => {
 		})
 		expect(sitemap.build()).toMatchSnapshot()
 	})
+})
 
+describe('image:image tests', () => {
+	it('should build corretcly with single image:image', () => {
+		const sitemap = new Sitemap('https://www.example.com')
+		sitemap.addEntry('/path', {
+			images: [{
+				location: '/test',
+				geoLocation: 'Nantes',
+				title: 'Title',
+				caption: 'Image Caption',
+				license: 'Example license url'
+			}]
+		})
+		expect(sitemap.build()).toMatchSnapshot()
+	})
+	it('should skip image:image if no location is set', () => {
+		const sitemap = new Sitemap('https://www.example.com')
+		sitemap.addEntry('/path', {
+			// @ts-expect-error
+			images: [{
+				geoLocation: 'Nantes',
+			}]
+		})
+		expect(sitemap.build()).toMatchSnapshot()
+	})
+	it('should skip image:image if there is more than 1000 images', () => {
+		const sitemap = new Sitemap('https://www.example.com')
+		sitemap.addEntry('/path', {
+			images: Array.from(new Array(1001)).map(() => ({
+				location: '/test',
+			}))
+		})
+		expect(sitemap.build()).toMatchSnapshot()
+	})
+	it('should build corretcly with multiple image:image', () => {
+		const sitemap = new Sitemap('https://www.example.com')
+		sitemap.addEntry('/path', {
+			images: [{
+				location: '/test',
+				geoLocation: 'Nantes',
+				title: 'Title',
+				caption: 'Image Caption',
+				license: 'Example license url'
+			}, {
+				location: '/test-2',
+				geoLocation: 'Paris',
+				title: 'Title2',
+				caption: 'Image Caption2',
+				license: 'Example license url'
+			}]
+		})
+		expect(sitemap.build()).toMatchSnapshot()
+	})
 })