+++ packages/dom-manager/CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2021-02-08
+
+### Added
+
+- function to add child .appendChild
+- Autocompletion for .style
+
+## [1.2.3] - 2020-12-17
+
+### Fixed
+
+- Setting .text to '' dont set text to nothing #6
+- t.data('test', null) return the value of test #7
+
 ## [1.2.2] - 2020-10-20
 
 ### Changed
@@ -38,6 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added base Manager
 
+[1.3.0]: https://github.com/dzeiocom/libs/releases/tag/%40dzeio%2Fdom-manager%401.3.0
+[1.2.2]: https://github.com/dzeiocom/libs/releases/tag/%40dzeio%2Fdom-manager%401.2.3
 [1.2.2]: https://github.com/dzeiocom/libs/releases/tag/%40dzeio%2Fdom-manager%401.2.2
 [1.2.1]: https://github.com/dzeiocom/libs/releases/tag/%40dzeio%2Fdom-manager%401.2.1
 [1.2.0]: https://github.com/dzeiocom/libs/releases/tag/%40dzeio%2Fdom-manager%401.2.0